backend/app/schemas/auth.py
from pydantic import BaseModel, Field


class LoginRequest(BaseModel):
    """登录请求"""

    username: str = Field(..., min_length=3, max_length=50)
    password: str = Field(..., min_length=6, max_length=100)


class LoginResponse(BaseModel):
    """登录响应"""

    access_token: str
    token_type: str = "bearer"
    username: str


class TokenPayload(BaseModel):
    """JWT Token 载荷"""

    sub: str  # username
    exp: int  # expiration time
