backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-dotenv==1.2.1
pydantic-settings==2.11.0
bcrypt==4.2.1
pyjwt==2.10.1
